Move useHelpLink into common/hooks
The hook reads the help-link settings, the current user and the
paid-plan status, so nothing about it is nav-specific. Moving it out
of nav/components/AppSwitcher removes the upsells module's only upward
import, so its enforceSharedTiers flag goes too.
