+++ frontend/src/metabase/common/components/upsells/StoragePurchaseModal/StorageSetupErrorView.tsx
@@ -1,7 +1,7 @@
 import { c, t } from "ttag";
 
 import { ExternalLink } from "metabase/common/components/ExternalLink";
-import { useHelpLink } from "metabase/nav/components/AppSwitcher/useHelpLink";
+import { useHelpLink } from "metabase/common/hooks";
 import { Icon } from "metabase/ui";
 
 import { StorageSetupStatusView } from "./StorageSetupStatusView";
