+++ frontend/src/metabase/nav/components/AppSwitcher/AppSwitcher.tsx
@@ -7,6 +7,7 @@ import { ExternalLink } from "metabase/common/components/ExternalLink";
 import { ForwardRefLink } from "metabase/common/components/Link";
 import { trackDataStudioOpened } from "metabase/common/data-studio/analytics";
 import { canAccessDataStudio as canAccessDataStudioSelector } from "metabase/common/data-studio/selectors";
+import { useHelpLink } from "metabase/common/hooks";
 import { trackMonitorOpened } from "metabase/common/monitor/analytics";
 import { canAccessMonitor as canAccessMonitorSelector } from "metabase/common/monitor/selectors";
 import { prepareInitials } from "metabase/common/utils/user";
@@ -40,7 +41,6 @@ import { AboutModal } from "../AboutModal/AboutModal";
 
 import S from "./AppSwitcher.module.css";
 import { useGetCurrentApp } from "./useGetCurrentApp";
-import { useHelpLink } from "./useHelpLink";
 
 const CURRENT_APP_ICON_OVERRIDES: {
   name: IconName;
